requirements.txt
# Streamlit and web framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Google Cloud and BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.0.0

# Network graph visualization
networkx>=3.1
pyvis>=0.3.2

# Environment and configuration
python-dotenv>=1.0.0

# Text processing and utilities
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: For advanced text processing
nltk>=3.8.1
spacy>=3.6.0
